+++ rules/batiment/batimentquadmaisonpont.cpp
@@ -1,6 +1,6 @@
 #include "all_includes.hh"
 
-BatimentQuadMaisonPont::BatimentQuadMaisonPont(Quad c, int height) : Chose(), c(c), height(height) {
+BatimentQuadMaisonPont::BatimentQuadMaisonPont(Quad c, float height) : Chose(), c(c), height(height) {
 	addEntropy(c);
 }
 
@@ -14,7 +14,7 @@ bool BatimentQuadMaisonPont::split() {
 	if(Segment(q[NE],q[NW]).length() < Segment(q[NE],q[SE]).length())
         q = q << 1;
     float partLength = Segment(q[NE],q[NW]).length() / 3;
-    int partHeight = 2.5*height/3.;
+    float partHeight = 2.5f * height / 3.f;
     Quad qa = q.inset(E,2*partLength);
     Quad qb = q.inset(W,2*partLength);
     Quad qc = q.inset(E, partLength).inset(W, partLength);
@@ -31,15 +31,15 @@ bool BatimentQuadMaisonPont::split() {
 }
 
 void BatimentQuadMaisonPont::triangulation() {
-	float h = 2.5*height/3.;
+	float h = 2.5f * height / 3.f;
 	Quad q = c.makeParallelogram();
 	Quad qh = q + Vertex(0,0,h);
 
     addGPUQuad(c,0x80,0x80,0x80);
     addGPUOcto(q,qh,0xF1,0xE0,0xE0);
 
-    Vertex ce = qh[SE] + (qh[NE] - qh[SE])/2 + Vertex(0,0,0.5*height/3.f);
-    Vertex cw = qh[SW] + (qh[NW] - qh[SW])/2 + Vertex(0,0,0.5*height/3.f);
+    Vertex ce = qh[SE] + (qh[NE] - qh[SE])/2 + Vertex(0,0,0.5f * height / 3.f);
+    Vertex cw = qh[SW] + (qh[NW] - qh[SW])/2 + Vertex(0,0,0.5f * height / 3.f);
 
     addGPUTriangle(qh[NW],cw,qh[SW],0xF1,0xE0,0xE0);
     addGPUTriangle(qh[SE],ce,qh[NE],0xF1,0xE0,0xE0);
